+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/application/StudentRestController.java
@@ -50,6 +50,10 @@ public class StudentRestController {
         return studentController.getStudentById(studentId);
     }
 
+    @DeleteMapping("/{studentId}")
+    public void deleteStudent(@PathVariable Long studentId) {
+        studentController.deleteStudent(studentId);
+    }