+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/controllers/StudentController.java
@@ -78,4 +78,7 @@ public class StudentController {
     }
 
 
+    public void deleteStudent(Long studentId) {
+        studentService.deleteStudent(studentId);
+    }
 }