+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/service/StudentService.java
@@ -31,4 +31,7 @@ public class StudentService {
     }
 
 
+    public void deleteStudent(Long studentId) {
+        studentRepository.deleteById(studentId);
+    }
 }