+++ NIMIO2024_TFG_PolMorenoCastelar/changelog.md
@@ -84,6 +84,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - [NPP-45] 》ST39F - Update student by ID
 - [NPP-48] 》ST42F - Get students by courseId
 - [NPP-49] 》ST43F - Get student by ID
+- [NPP-54] 》ST48F - Delete a student by ID
 
     ## Student Exceptions